Move C code to its own file, to remove dependency on RubyInline
